Setup Window System
Created a way to display windows from a single panel and prefabs.
Currently creates a clone of the prefab when the function executes, and
then finds and removes it on the close function. One issue that will
need to be resolved is that other UI elements will still be functional
with a window open. Hopefully some boolean variable is all that is
needed. added the openwindow and closewindow commands to the console
along with the only window currently: devTroopWindow.