Window System work, Troop Calculation Creation
More tweaks to the window system, such as preventing things from working
when a window is open. Also setup the system for calculating battle
mechanics., all that's left is to work on actual calculations. Still
need to get it to show the remaining troops, but that will take some
recoding, refer to the devTroopWindow.cs comment for details.